Add files via upload
A pharmacy management system is a system that consists of data entry, retrieval and monitoring stock, sale, customer records, debtor’s and management administrator’s records, and determination of the minimum quantity of each drug. String searching technique also applied in this system. This technique is referring to drug name, drug code, and description of drugs. Besides that, the system also provides two types of methods which are Quantity and Expire date of drugs. This system always checking the date to remind the salesman if the certain drug was expired and will be triggered to remind the salesman if certain of the drugs reached the minimum quantity. This system enables the administrator to control and monitor the drug stock effectively.
Functional Requirements: -
  The function what the system does is  
⦁	Store Medicine data’s
⦁	Search Medicine data effectively
⦁	Update, delete, and Edit medicine information
⦁	Generate report on medicine
⦁	Prepare bill for the medicine
⦁	Give alert for the user when medicine item is low
⦁	Changing login password